// src/module/entities/model/ConceptsToProduct.ts
export type ConceptToProduct = {
  sigemp: string;
  situacao: number;
  conceitoCod: number;
  produtoCod: number;
  dtAlteracao: Date;
  hrAlteracao: number;
};
export type ConceptToProductFields = {
  sigemp?: boolean;
  situacao?: boolean;
  conceitoCod?: boolean;
  produtoCod?: boolean;
  dtAlteracao?: boolean;
  hrAlteracao?: boolean;
};